requirements.txt
django-admin-honeypot
django-allauth
django-appconf
django-bootstrap-form
django-bootstrap4
django-braces
django-celery-beat
django-channels
django-ckeditor
django-cleanup
django-compressor
django-cors-headers
django-crispy-forms
django-debug-toolbar
django-encrypted-model-fields
django-environ
django-extensions
django-filter
django-flatpickr
django-graphql-jwt
django-haystack
django-heroku
django-hosts
django-js-asset
django-markdown
django-model-utils
django-paypal
django-phonenumber-field
django-private-chat
django-redis
django-request
django-safedelete
django-softdelete
django-storages
django-tables2
django-taggit
django-timezone-field
django-urlcrypt
django-weasyprint
django-widget-tweaks
djangorestframework
djangorestframework-jwt